[-] 폴더 구조 변경 / SIgnUp, SignIn Page 수정 / 주석 변경
