--- client/src/components/auth/SignIn/index.js
@@ -1,2 +1,0 @@
-export { default as SignInForm } from './SignInForm';
-export { default as SignIn } from './SignIn';
