+++ client/src/components/auth/SignInForm.jsx
@@ -5,10 +5,10 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useNavigate } from 'react-router-dom';
 import { useSetRecoilState } from 'recoil';
 import { Stack, Input, Button } from '@mantine/core';
-import { signIn } from '../../../api/auth';
-import { InputWrapper } from '../../common/form';
-import userState from '../../../recoil/atoms/userState';
-import routesConstants from '../../../constants/routes';
+import { signIn } from '../../api/auth';
+import { InputWrapper } from '../common/form';
+import userState from '../../recoil/atoms/userState';
+import routesConstants from '../../constants/routes';
 
 const signinScheme = z.object({
   email: z.string().email({ message: '이메일 형식에 맞게 입력해 주세요.' }),
