--- client/src/components/auth/SignUp/SignUp.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Container, Title } from '@mantine/core';
-import { SignUpForm } from '.';
-
-const SignUp = () => (
-  <Container c="var(--font-color)" py="48px" size="xs">
-    <Title mb="xl" align="center">
-      FineApple ID 생성
-    </Title>
-    <SignUpForm />
-  </Container>
-);
-export default SignUp;
