--- client/src/components/auth/SignUp/index.js
@@ -1,3 +1,0 @@
-export { default as PasswordTooltipInput } from './PasswordTooltipInput';
-export { default as SignUpForm } from './SignUpForm';
-export { default as SignUp } from './SignUp';
