+++ client/src/components/auth/SignUpForm.jsx
@@ -4,11 +4,11 @@ import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useNavigate } from 'react-router-dom';
 import { Stack, Flex, Button, Input, Divider } from '@mantine/core';
-import { checkEmail, checkNickName, signUp } from '../../../api/auth';
+import { checkEmail, checkNickName, signUp } from '../../api/auth';
 import { PasswordTooltipInput } from '.';
-import { InputWrapper, CountrySelect, BirthDateInput, DuplicateCheckInput, PhoneNumberInput } from '../../common/form';
-import routesConstants from '../../../constants/routes';
-import useToast from '../../../hooks/useToast';
+import { InputWrapper, CountrySelect, BirthDateInput, DuplicateCheckInput, PhoneNumberInput } from '../common/form';
+import routesConstants from '../../constants/routes';
+import useToast from '../../hooks/useToast';
 
 const signupScheme = z
   .object({
