+++ client/src/components/auth/index.js
@@ -1,0 +1,3 @@
+export { default as SignUpForm } from './SignUpForm';
+export { default as SignInForm } from './SignInForm';
+export { default as PasswordTooltipInput } from './PasswordTooltipInput';
